style: satisfy the CI clippy gate
`just ci`'s lint leg is stricter than a plain `cargo clippy`: an unnecessary
`f64 -> f64` cast in the new `Switch` track radius, and a `Default::default()`
followed by field assignment in the colour picker's readout swatch.

+++ crates/lumen-widgets/src/controls.rs
@@ -456,7 +456,7 @@ impl Switch {
                 } else {
                     Color::srgb8(0xcc, 0xcc, 0xcc, 0xff)
                 }),
-                corner_radius: TRACK_H as f64 / 2.0,
+                corner_radius: TRACK_H / 2.0,
                 style: LayoutStyle {
                     position: Position::Relative,
                     width: Dim::px(TRACK_W as f32),
